requirements.txt
fastapi
openai
python-dotenv
pydantic
googleapiclient
youtube_transcript_api
spacy
tiktoken